test: separate directory enricher coverage
